+++ src/quacc/recipes/gulp/core.py
@@ -7,9 +7,9 @@ from typing import TYPE_CHECKING
 from ase.calculators.gulp import GULP
 
 from quacc import SETTINGS, job
+from quacc.runners.calc import run_calc
 from quacc.schemas import fetch_atoms
 from quacc.schemas.ase import summarize_run
-from quacc.utils.calc import run_calc
 from quacc.utils.dicts import merge_dicts
 
 if TYPE_CHECKING: