+++ src/quacc/recipes/psi4/core.py
@@ -7,9 +7,9 @@ from ase.calculators.psi4 import Psi4
 from monty.dev import requires
 
 from quacc import job
+from quacc.runners.calc import run_calc
 from quacc.schemas import fetch_atoms
 from quacc.schemas.ase import summarize_run
-from quacc.utils.calc import run_calc
 from quacc.utils.dicts import merge_dicts
 
 try: