+++ src/quacc/recipes/tblite/core.py
@@ -7,6 +7,8 @@ from ase.optimize import FIRE
 from monty.dev import requires
 
 from quacc import job
+from quacc.runners.calc import run_ase_opt, run_ase_vib, run_calc
+from quacc.runners.thermo import ideal_gas
 from quacc.schemas import fetch_atoms
 from quacc.schemas.ase import (
     summarize_opt_run,
@@ -14,9 +16,7 @@ from quacc.schemas.ase import (
     summarize_thermo,
     summarize_vib_run,
 )
-from quacc.utils.calc import run_ase_opt, run_ase_vib, run_calc
 from quacc.utils.dicts import merge_dicts
-from quacc.utils.thermo import ideal_gas
 
 try:
     from tblite.ase import TBLite
@@ -120,7 +120,7 @@ def relax_job(
     calc_swaps
         Dictionary of custom kwargs for the tblite calculator.
     opt_swaps
-        Dictionary of custom kwargs for [quacc.utils.calc.run_ase_opt][].
+        Dictionary of custom kwargs for [quacc.runners.calc.run_ase_opt][].
     copy_files
         Files to copy to the runtime directory.