+++ src/quacc/recipes/vasp/qmof.py
@@ -13,9 +13,9 @@ from ase.optimize import BFGSLineSearch
 from quacc import job
 from quacc.calculators.vasp import Vasp
 from quacc.recipes.vasp.core import _base_job
+from quacc.runners.calc import run_ase_opt
 from quacc.schemas import fetch_atoms
 from quacc.schemas.ase import summarize_opt_run
-from quacc.utils.calc import run_ase_opt
 from quacc.utils.dicts import merge_dicts
 
 if TYPE_CHECKING: