+++ src/quacc/schemas/vasp.py
@@ -13,11 +13,11 @@ from pymatgen.command_line.bader_caller import bader_analysis_from_path
 from pymatgen.command_line.chargemol_caller import ChargemolAnalysis
 
 from quacc import SETTINGS
+from quacc.runners.prep import prep_next_run as prep_next_run_
 from quacc.schemas.atoms import atoms_to_metadata
-from quacc.utils.atoms import prep_next_run as prep_next_run_
-from quacc.utils.db import results_to_db
 from quacc.utils.dicts import clean_dict
 from quacc.utils.files import copy_decompress
+from quacc.wflow.db import results_to_db
 
 if TYPE_CHECKING:
     from typing import TypeVar