+++ CHANGELOG.md
@@ -165,7 +165,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Added support for Prefect.
-- Added `quacc.utils.wflows` module.
+- Added `quacc.wflow` module.
 
 ### Changed